requirements.txt
astroid==1.5.3
Babel==2.4.0
certifi==2017.4.17
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.1
coreschema==0.0.4
Django==1.11.2
django-admin-bootstrapped==2.5.7
django-authtools==1.6.0
django-braces==1.11.0
django-crispy-forms==1.6.1
django-debug-toolbar==1.8
django-environ==0.4.3
django-extensions==1.7.9
django-extra-fields==0.9
django-filter==1.0.4
django-phonenumber-field==1.3.0
djangorestframework==3.6.3
easy-thumbnails==2.4.1
idna==2.5
isort==4.2.15
itypes==1.1.0
Jinja2==2.9.6
lazy-object-proxy==1.3.1
Markdown==2.6.8
MarkupSafe==1.0
mccabe==0.6.1
olefile==0.44
phonenumberslite==8.5.1
Pillow==4.1.1
psycopg2==2.7.1
PyJWT==1.5.0
pylint==1.7.1
pylint-django==0.7.2
pylint-plugin-utils==0.2.6
pytz==2017.2
requests==2.18.1
six==1.10.0
sqlparse==0.2.3
uritemplate==3.0.0
urllib3==1.21.1
Werkzeug==0.12.2
wrapt==1.10.10
